# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/collections/{id}/products/{productId}": {
      "post": {
        "operationId": "AdminCollectionsController_addProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a product to a collection (Admin)",
        "tags": [
          "admin collections"
        ]
      },
      "delete": {
        "operationId": "AdminCollectionsController_removeProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a product from a collection (Admin)",
        "tags": [
          "admin collections"
        ]
      }
    },
...
    "/api/v1/products/{slug}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active product returned."
          },
          "404": {
            "description": "Product not found."
          }
        },
        "summary": "Get a single product with full details",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/products/reorder": {
      "patch": {
        "operationId": "ProductsController_reorder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderProductsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products reordered successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reorder products manually (Admin)",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully."
          },
          "404": {
            "description": "Product not found."
          },
          "409": {
            "description": "A product with this slug already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product details (Admin)",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product archived successfully."
          },
          "404": {
            "description": "Product not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a product without deleting order history (Admin)",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/admin/products/reorder": {
      "patch": {
        "operationId": "AdminProductsController_reorder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderProductsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products reordered successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reorder products manually (Admin)",
        "tags": [
          "admin products"
        ]
      }
    },
    "/api/v1/admin/products": {
      "get": {
        "operationId": "AdminProductsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "ACTIVE",
                "PUBLISHED",
                "ARCHIVED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated products returned.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "meta": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List products in every status (Admin)",
        "tags": [
          "admin products"
        ]
      }
    },
    "/api/v1/admin/products/{id}": {
      "get": {
        "operationId": "AdminProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product returned successfully."
          },
          "404": {
            "description": "Product not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a product in any status by id (Admin)",
        "tags": [
          "admin products"
        ]
      }
    },
    "/api/v1/admin/products/{productId}/variants": {
      "post": {
        "operationId": "AdminProductsController_addVariant",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductVariantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Variant created successfully."
          },
          "404": {
            "description": "Product not found."
          },
          "409": {
            "description": "A variant with this SKU already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a variant to a product (Admin)",
        "tags": [
          "admin products"
        ]
      }
    },
    "/api/v1/admin/products/{productId}/variants/{variantId}": {
      "patch": {
        "operationId": "AdminProductsController_updateVariant",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductVariantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Variant updated successfully."
          },
          "404": {
            "description": "Product or variant not found."
          },
          "409": {
            "description": "Variant was modified concurrently, or SKU conflict."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Partially update a product variant (stock/price/availability/color/size) (Admin)",
        "tags": [
          "admin products"
        ]
      },
      "delete": {
        "operationId": "AdminProductsController_removeVariant",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Variant deleted successfully."
          },
          "404": {
            "description": "Product or variant not found."
          },
          "409": {
            "description": "Variant has existing order items and cannot be deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a product variant that has no order history (Admin)",
        "tags": [
          "admin products"
        ]
      }
    },
    "/api/v1/admin/products/{productId}/images": {
      "post": {
        "operationId": "AdminProductsController_addImage",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductImageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image created successfully."
          },
          "404": {
            "description": "Product not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an image to a product (Admin)",
        "tags": [
          "admin products"
        ]
      }
    },
    "/api/v1/admin/products/{productId}/images/{imageId}": {
      "delete": {
        "operationId": "AdminProductsController_removeImage",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image deleted successfully."
          },
          "404": {
            "description": "Product or image not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a product image (Admin)",
        "tags": [
          "admin products"
        ]
      }
    },
...
    "/api/v1/settings/site": {
      "get": {
        "operationId": "SiteSettingsController_getSiteSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Site settings returned."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get site settings (Admin)",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "SiteSettingsController_updateSiteSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSiteSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Site settings updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update site settings (Admin)",
        "tags": [
          "settings"
        ]
      }
    },
...
  },
  "info": {
    "title": "RIVE Luxury Store API",
    "description": "Luxury e-commerce backend for the RIVE storefront",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReorderProductsDto": {
        "type": "object",
        "properties": {
          "productIds": {
            "example": [
              "prod_1",
              "prod_2",
              "prod_3"
            ],
            "description": "Ordered array of product IDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "productIds"
        ]
      },
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 256
          },
          "slug": {
            "type": "string",
            "maxLength": 256
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "shortDescription": {
            "type": "string",
            "maxLength": 500
          },
          "seoTitle": {
            "type": "string",
            "maxLength": 256
          },
          "seoDescription": {
            "type": "string",
            "maxLength": 500
          },
          "displayOrder": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "compareAtPrice": {
            "type": "number"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "PUBLISHED",
              "ARCHIVED"
            ]
          }
        }
      },
      "CreateProductVariantDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "example": "LUNA-SET-S-BEIGE",
            "maxLength": 128
          },
          "colorHex": {
            "type": "string",
            "example": "#945958",
            "default": "#945958"
          },
          "size": {
            "type": "string",
            "enum": [
              "XS",
              "S",
              "M",
              "L",
              "XL"
            ],
            "example": "S"
          },
          "price": {
            "type": "number",
            "example": 280
          },
          "stock": {
            "type": "number",
            "example": 12,
            "default": 0
          },
          "isAvailable": {
            "type": "boolean",
            "example": true,
            "default": true
          }
        },
        "required": [
          "sku",
          "size",
          "price"
        ]
      },
      "UpdateProductVariantDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "example": "LUNA-SET-S-BEIGE",
            "maxLength": 128
          },
          "colorHex": {
            "type": "string",
            "example": "#945958"
          },
          "size": {
            "type": "string",
            "enum": [
              "XS",
              "S",
              "M",
              "L",
              "XL"
            ],
            "example": "S"
          },
          "price": {
            "type": "number",
            "example": 280
          },
          "stock": {
            "type": "number",
            "example": 12,
            "description": "Absolute stock value to set. Not a delta/decrement — actual sale-time decrements happen atomically in orders.service.ts."
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "expectedUpdatedAt": {
            "type": "string",
            "example": "2026-09-03T12:00:00.000Z",
            "description": "The variant's current updatedAt, as returned by a prior GET/create/update response. Used for optimistic-locking: the update is rejected with 409 Conflict if the variant was modified by someone else in the meantime."
          }
        },
        "required": [
          "expectedUpdatedAt"
        ]
      },
      "CreateProductImageDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://images.example.com/product-1.png",
            "maxLength": 2048
          },
          "altText": {
            "type": "string",
            "example": "Luna silk set front view",
            "maxLength": 500
          },
          "isPrimary": {
            "type": "boolean",
            "example": true,
            "default": false
          }
        },
        "required": [
          "url"
        ]
      },
...
      "UpdateSiteSettingsDto": {
        "type": "object",
        "properties": {
          "storeName": {
            "type": "string",
            "example": "RIVÉ Store"
          },
          "logoUrl": {
            "type": "string",
            "example": "https://example.com/logo.png"
          },
          "phone": {
            "type": "string",
            "example": "+201000000000"
          },
          "email": {
            "type": "string",
            "example": "contact@rive.com"
          },
          "address": {
            "type": "string",
            "example": "Cairo, Egypt"
          },
          "facebookUrl": {
            "type": "string",
            "example": "https://facebook.com/rive"
          },
          "instagramUrl": {
            "type": "string",
            "example": "https://instagram.com/rive"
          },
          "whatsappNumber": {
            "type": "string",
            "example": "+201000000000"
          },
          "tiktokUrl": {
            "type": "string",
            "example": "https://tiktok.com/@rive"
          },
          "isMaintenanceMode": {
            "type": "boolean",
            "example": false
          },
          "minimumOrderAmount": {
            "type": "number",
            "example": 100
          },
          "businessHours": {
            "type": "object",
            "example": {
              "sunday": {
                "open": "09:00",
                "close": "22:00"
              }
            }
          }
        }
      },
...
    }
  }
}